third_party/blink/web_tests/fast/hidpi/static/mousewheel-scroll-amount.html

<!DOCTYPE html>
<script src="../../../resources/gesture-util.js"></script>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>

<style type="text/css">
    body {
        height: 2000px;
        width: 2000px;
    }
</style>

<script>
async function reset() {
    await waitForScrollReset(document.scrollingElement);
    await waitForCompositorCommit();
}

async function waitForWheelTickAndScroll(xTicks, yTicks, position) {
  const scrollPromise = waitForScrollendEvent(document);
  const gesturePromise = wheelTick(xTicks, yTicks, position);
  return Promise.all([gesturePromise, scrollPromise]);
}

var position = {x: 20, y: 20};
promise_test(async () => {
    await reset();
    await waitForWheelTickAndScroll(0, 1, position);
    const expectedScrollDeltaY = pixelsPerTick();
    if (internals.runtimeFlags.fractionalScrollOffsetsEnabled) {
        assert_approx_equals(document.scrollingElement.scrollLeft, 0, 0.0001);
        assert_approx_equals(document.scrollingElement.scrollTop, expectedScrollDeltaY, 0.0001);
    } else {
        assert_equals(Math.round(document.scrollingElement.scrollLeft), 0);
        assert_equals(Math.round(document.scrollingElement.scrollTop), expectedScrollDeltaY);
    }
}, "One vertical wheel tick scrolls the right number of pixels in high dpi mode.");

promise_test(async () => {
    await reset();
    await waitForWheelTickAndScroll(1, 0, position);
    const expectedScrollDeltaX = pixelsPerTick();
    if (internals.runtimeFlags.fractionalScrollOffsetsEnabled) {
        assert_approx_equals(document.scrollingElement.scrollLeft, expectedScrollDeltaX, 0.0001);
        assert_approx_equals(document.scrollingElement.scrollTop, 0, 0.0001);
    } else {
        assert_equals(Math.round(document.scrollingElement.scrollLeft), expectedScrollDeltaX);
        assert_equals(Math.round(document.scrollingElement.scrollTop), 0);
    }
}, "One horizontal wheel tick scrolls the right number of pixels in high dpi mode.");

promise_test(async () => {
    await reset();
    await waitForWheelTickAndScroll(0, 2, position);
    const expectedScrollDeltaY = pixelsPerTick() * 2;
    if (internals.runtimeFlags.fractionalScrollOffsetsEnabled) {
        assert_approx_equals(document.scrollingElement.scrollLeft, 0, 0.0001);
        assert_approx_equals(document.scrollingElement.scrollTop, expectedScrollDeltaY, 0.0001);
    } else {
        assert_equals(Math.round(document.scrollingElement.scrollLeft), 0);
        assert_equals(Math.round(document.scrollingElement.scrollTop), expectedScrollDeltaY);
    }
}, "Two vertical wheel ticks scroll the right number of pixels in high dpi mode.");

promise_test(async () => {
    await reset();
    await waitForWheelTickAndScroll(2, 0, position);
    const expectedScrollDeltaX = pixelsPerTick() * 2;
    if (internals.runtimeFlags.fractionalScrollOffsetsEnabled) {
        assert_approx_equals(document.scrollingElement.scrollLeft, expectedScrollDeltaX, 0.0001);
        assert_approx_equals(document.scrollingElement.scrollTop, 0, 0.0001);
    } else {
        assert_equals(Math.round(document.scrollingElement.scrollLeft), expectedScrollDeltaX);
        assert_equals(Math.round(document.scrollingElement.scrollTop), 0);
    }
}, "Two horizontal wheel ticks scroll the right number of pixels in high dpi mode.");
</script>
